made DomReader behave more like a fluent-iface
prev. all moveTo... methods has a mandatory argument and returned a boolean,
this had been split into two methods for each prev. moveTo... method. The one
keeping the name does not have a mandatory arg as moveTo... throws an exception
if the action can not be completed and returns 'this'. And another one named
tryMoveTo... which does not throw and returns a boolean.